+++ packages/frontend/synthesizer/core/src/synthesizer/handlers/arithmeticManager.ts
@@ -129,14 +129,6 @@ export class ArithmeticManager {
       }
     }
 
-    if (name === 'SIGNEXTEND') {
-      if (inPts[0] !== undefined && inPts[0].value > 31n) {
-        throw new Error(
-          `Synthesizer: Operation ${name} has an index or size value greater than 31. Adjust ${subcircuitName} subcircuit in qap-compiler.`,
-        );
-      }
-    }
-
     return { subcircuitName, finalInPts };
   }